ANDROID: modpost: Forbid double exports
In the context of GKI we are exporting quite a few internal kernel
symbols, some of which may be static functions (yes this is bad, but
technically it works). As such, we need to be careful about conflicts
and we really should not allow a symbol to be exported more than once.

To ensure this is the case, turn the modpost check 'fatal' to bail out
immediately when such a situation is detected.

Bug: 174214891
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I674c9c9c327849503e94d79654e6288411a830af
Signed-off-by: Matthias Maennich <maennich@google.com>